code added

+++ src/pages/dailyTrips/TripList.tsx
@@ -7,9 +7,11 @@ import {
   query,
   orderBy,
 } from "firebase/firestore";
-import { db } from "../../firebaseConfig";
+import { db } from "../../firebaseConfigTest";
 import { useNavigate } from "react-router-dom";
 import downloadBiltyPDF from "./generatepdf1";
+import { formatDateMMDDYYYY } from "../../utils";
+import DeleteConfirmModal from "../../components/DeleteConfirmModal";
 
 export default function TripList() {
   const navigate = useNavigate();
@@ -18,11 +20,26 @@ export default function TripList() {
   const [customers, setCustomers] = useState([]);
   const [drivers, setDrivers] = useState([]);
 
+  const today = new Date().toISOString().substring(0, 10);
+
   // Filters
   const [filterCustomer, setFilterCustomer] = useState("");
+  const [localCustomerName, setLocalCustomerName] = useState("");
+
+  const [showDeleteModal, setShowDeleteModal] = useState(false);
+  const [deleteTripId, setDeleteTripId] = useState(null);
+
   const [filterDriver, setFilterDriver] = useState("");
   const [filterLorryNo, setFilterLorryNo] = useState("");
   const [filterDate, setFilterDate] = useState("");
+
+  // Pagination
+  const PAGE_SIZE = 5;
+  const [currentPage, setCurrentPage] = useState(1);
+
+  // Date range filters
+  const [fromDate, setFromDate] = useState("");
+  const [toDate, setToDate] = useState("");
 
   useEffect(() => {
     loadTrips();
@@ -30,8 +47,17 @@ export default function TripList() {
     loadDrivers();
   }, []);
 
-  const loadTrips = async () => {
+  /* const loadTrips = async () => {
     const q = query(collection(db, "dailyTrips"), orderBy("lorryNo", "desc"));
+    const snap = await getDocs(q);
+    setTrips(snap.docs.map((d) => ({ id: d.id, ...d.data() })));
+  };*/
+  const loadTrips = async () => {
+    const q = query(
+      collection(db, "dailyTrips"),
+      orderBy("addTripDate", "desc") // 🔥 latest first
+    );
+
     const snap = await getDocs(q);
     setTrips(snap.docs.map((d) => ({ id: d.id, ...d.data() })));
   };
@@ -47,19 +73,80 @@ export default function TripList() {
   };
 
   const deleteTrip = async (id) => {
+    const role = localStorage.getItem("role");
+    if (role !== "1") {
+      alert("Only Admin can delete Trips");
+      return;
+    }
     await deleteDoc(doc(db, "dailyTrips", id));
     setTrips(trips.filter((t) => t.id !== id));
   };
 
+  const handleDeleteClick = (id) => {
+    const role = localStorage.getItem("role");
+    setDeleteTripId(id);
+    setShowDeleteModal(true);
+  };
+
+  const confirmDeleteTrip = async () => {
+    try {
+      await deleteDoc(doc(db, "dailyTrips", deleteTripId));
+      setTrips((prev) => prev.filter((t) => t.id !== deleteTripId));
+    } catch (error) {
+      console.error("Delete failed", error);
+      alert("Failed to delete trip");
+    } finally {
+      setShowDeleteModal(false);
+      setDeleteTripId(null);
+    }
+  };
+
+  const cancelDelete = () => {
+    setShowDeleteModal(false);
+    setDeleteTripId(null);
+  };
+
   // Filtering logic
   const filteredTrips = trips.filter((t) => {
+    const tripDate = t.addTripDate?.substring(0, 10); // yyyy-mm-dd
+
+    console.log("Trip Date:", tripDate, "Filter Date:", filterDate);
+    const isWithinDateRange =
+      (!fromDate || tripDate >= fromDate) && (!toDate || tripDate <= toDate);
+
     return (
       (filterCustomer ? t.customerId === filterCustomer : true) &&
       (filterDriver ? t.driverId === filterDriver : true) &&
-      (filterLorryNo ? t.lorryNo.toString().includes(filterLorryNo) : true) &&
-      (filterDate ? t.addedDate?.substring(0, 10) === filterDate : true)
+      (filterLorryNo ? t.lorryNo?.toString().includes(filterLorryNo) : true) &&
+      (localCustomerName
+        ? t.localCustomerName?.toString().includes(localCustomerName)
+        : true) &&
+      isWithinDateRange
     );
   });
+
+  console.log("Filtered Trips:", filteredTrips);
+
+  const totalPages = Math.ceil(filteredTrips.length / PAGE_SIZE);
+
+  const paginatedTrips = filteredTrips.slice(
+    (currentPage - 1) * PAGE_SIZE,
+    currentPage * PAGE_SIZE
+  );
+
+  useEffect(() => {
+    setCurrentPage(1);
+  }, [filterCustomer, filterDriver, filterLorryNo, fromDate, toDate]);
+
+  const clearFilters = () => {
+    setFilterCustomer("");
+    setLocalCustomerName("");
+    setFilterDriver("");
+    setFilterLorryNo("");
+    setFromDate("");
+    setToDate("");
+    setCurrentPage(1);
+  };
 
   return (
     <div className="page-container">
@@ -68,6 +155,8 @@ export default function TripList() {
       <button onClick={() => navigate("/add-daily-trip")}>
         + Add Daily Trip
       </button>
+
+      <button onClick={() => clearFilters()}>Clear Filters</button>
 
       {/* FILTERS */}
       <div
@@ -85,6 +174,13 @@ export default function TripList() {
             </option>
           ))}
         </select>
+
+        <input
+          type="text"
+          placeholder="Local Customer"
+          value={localCustomerName}
+          onChange={(e) => setLocalCustomerName(e.target.value)}
+        />
 
         <select
           value={filterDriver}
@@ -104,11 +200,23 @@ export default function TripList() {
           value={filterLorryNo}
           onChange={(e) => setFilterLorryNo(e.target.value)}
         />
+      </div>
+      <div
+        className="filters"
+        style={{ display: "flex", gap: "10px", marginTop: "20px" }}
+      >
+        <input
+          type="date"
+          value={fromDate}
+          onChange={(e) => setFromDate(e.target.value)}
+          placeholder="From date"
+        />
 
         <input
           type="date"
-          value={filterDate}
-          onChange={(e) => setFilterDate(e.target.value)}
+          value={toDate}
+          onChange={(e) => setToDate(e.target.value)}
+          placeholder="To date"
         />
       </div>
 
@@ -122,34 +230,33 @@ export default function TripList() {
             <th>From</th>
             <th>To</th>
             <th>Invoice Value</th>
-            <th>Bilty</th>
-            <th>Date</th>
+            <th>Trip Date</th>
+            <th>Added Date</th>
             <th>Actions</th>
           </tr>
         </thead>
 
         <tbody>
-          {filteredTrips.map((t) => {
+          {paginatedTrips.map((t) => {
             const customer = customers.find((c) => c.id === t.customerId);
             const driver = drivers.find((d) => d.id === t.driverId);
 
             return (
               <tr key={t.id}>
-                <td>{t.lorryNo}</td>
-                <td>{customer?.name}</td>
+                <td>{t.lorryNo ? t.lorryNo : "N/A"}</td>
+                <td>
+                  {t?.localCustomerName ? t?.localCustomerName : customer?.name}{" "}
+                </td>
                 <td>{driver?.name}</td>
                 <td>{t.fromLocation}</td>
                 <td>{t.toLocation}</td>
-                <td>₹ {t.invoiceValue}</td>
+                <td>₹ {t.tripCharges}</td>
 
                 {/* BILTY BUTTON */}
-
-                <td>{t.addedDate}</td>
+                <td>{formatDateMMDDYYYY(t.addTripDate)}</td>
+                <td>{formatDateMMDDYYYY(t.addedDate)}</td>
 
                 <td>
-                  <button onClick={() => downloadBiltyPDF(t)}>
-                    Download Bilty
-                  </button>
                   <button onClick={() => navigate(`/edit-daily-trip/${t.id}`)}>
                     Edit
                   </button>
@@ -159,7 +266,7 @@ export default function TripList() {
                       background: "red",
                       color: "#fff",
                     }}
-                    onClick={() => deleteTrip(t.id)}
+                    onClick={() => handleDeleteClick(t.id)}
                   >
                     Delete
                   </button>
@@ -175,8 +282,40 @@ export default function TripList() {
               </td>
             </tr>
           )}
+
+          <tr>
+            <td colSpan={10} style={{ textAlign: "center", padding: "20px" }}>
+              <div style={{ marginTop: "20px", textAlign: "center" }}>
+                <button
+                  disabled={currentPage === 1}
+                  onClick={() => setCurrentPage((p) => p - 1)}
+                >
+                  Prev
+                </button>
+
+                <span style={{ margin: "0 10px" }}>
+                  Page {currentPage} of {totalPages}
+                </span>
+
+                <button
+                  disabled={currentPage === totalPages}
+                  onClick={() => setCurrentPage((p) => p + 1)}
+                >
+                  Next
+                </button>
+              </div>
+            </td>
+          </tr>
         </tbody>
       </table>
+
+      <DeleteConfirmModal
+        open={showDeleteModal}
+        title="Delete Trip"
+        message="Are you sure you want to delete this trip? This action cannot be undone."
+        onConfirm={confirmDeleteTrip}
+        onCancel={cancelDelete}
+      />
     </div>
   );
 }
